% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{check_candidate_ids}
\alias{check_candidate_ids}
\title{Check if candidate IDs are ready for analysis}
\usage{
check_candidate_ids(idlist, datadir, required_files)
}
\arguments{
\item{idlist}{<what param does>}

\item{datadir}{<what param does>}

\item{required_files}{<what param does>}
}
\description{
Check if candidate IDs are ready for analysis
}
