% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_analyses.r
\name{determine_analyses}
\alias{determine_analyses}
\title{What analyses are to be run?}
\usage{
determine_analyses(id, idlist, newidlist = NULL, what = "eve")
}
\arguments{
\item{id}{target ID being performed by this node}

\item{idlist}{List of IDs in dataset}

\item{newidlist}{Default = NULL. If this is updating an existing analysis then need to use idlist for OLD ids, and newidlist for NEW ids}

\item{what="eve"}{Can be "phewas", "exposure", "outcome", "eve"}
}
\value{
data frame
}
\description{
For analysis of an ID in terms of PheWAS, just exposure or just outcome is straightforward. However, to either create or update everything-versus-everything we need to make sure we are not duplicating analyses across parallel processes. Strategy is to always just to exposure, because all other ids will perform the outcome analyses. However, when updating an existing graph it gets a bit more complicated - need to also make sure that both exposure and outcome analyses for new ID's are being done without duplicating analysis of old ID's
}
