% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{initialise_analysis}
\alias{initialise_analysis}
\title{Generate viable candidate analysis list}
\usage{
initialise_analysis(candidate, description)
}
\arguments{
\item{candidate}{trait_a, trait_b, method}
}
\description{
Generate viable candidate analysis list
}
