% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j.r
\name{modify_node_headers_for_neo4j}
\alias{modify_node_headers_for_neo4j}
\title{Modify headers for neo4j}
\usage{
modify_node_headers_for_neo4j(x, id, idname)
}
\arguments{
\item{x}{<what param does>}

\item{id}{<what param does>}

\item{idname}{<what param does>}
}
\description{
Modify headers for neo4j
}
