% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j.r
\name{modify_rel_headers_for_neo4j}
\alias{modify_rel_headers_for_neo4j}
\title{Modify headers for neo4j}
\usage{
modify_rel_headers_for_neo4j(x, id1, id1name, id2, id2name)
}
\arguments{
\item{x}{<what param does>}

\item{id1}{<what param does>}

\item{id1name}{<what param does>}

\item{id2}{<what param does>}

\item{id2name}{<what param does>}
}
\description{
<full description>
}
