% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{readml}
\alias{readml}
\title{Read in extracted data format}
\usage{
readml(filename, ao, format = "TwoSampleMR")
}
\arguments{
\item{filename}{filename e.g. ml.csv.gz}
}
\description{
Read in extracted data format
}
