% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.r
\name{setup_inventory}
\alias{setup_inventory}
\title{Initial creation of inventory}
\usage{
setup_inventory(path = NULL)
}
\arguments{
\item{path=NULL}{<what param does>}
}
\description{
Initial creation of inventory
}
