% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j.r
\name{write_simple}
\alias{write_simple}
\title{Wrapper to write out files}
\usage{
write_simple(obj, filename, id1, id1name, id2 = NULL, id2name = NULL)
}
\arguments{
\item{obj}{<what param does>}

\item{filename}{<what param does>}

\item{id1}{<what param does>}

\item{id1name}{<what param does>}

\item{id2=NULL}{<what param does>}

\item{id2name=NULL}{<what param does>}
}
\description{
<full description>
}
