% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j.r
\name{write_split}
\alias{write_split}
\title{Write out to csv.gz, split to make more manageable files}
\usage{
write_split(obj, splitsize, prefix, id1, id1name, id2 = NULL,
  id2name = NULL)
}
\arguments{
\item{obj}{<what param does>}

\item{splitsize}{<what param does>}

\item{prefix}{<what param does>}

\item{id1}{<what param does>}

\item{id1name}{<what param does>}

\item{id2=NULL}{<what param does>}

\item{id2name=NULL}{<what param does>}
}
\description{
Write out to csv.gz, split to make more manageable files
}
